Support double uiscale for GameShell(1.25)

+++ src/system.h
@@ -81,7 +81,7 @@ typedef struct
 	const char* crtShader;
 	const tic_cartridge* cart;
 
-	s32 uiScale;
+	double uiScale;
 
 } StudioConfig;
